Fixed bug with float instead of double.

+++ draw/path.h
@@ -39,7 +39,7 @@ public:
 	void cubic_to( const point &p1, const point &p2, const point &p3 );
 //	virtual void arc_to( const point &p1, const &point &p2, double r );
 	void arc_to( const point &center, const point &radius, double angle1, double angle2 );
-	void arc_to( const point &center, float radius, double angle1, double angle2 ) { arc_to( center, { center.x() + radius, center.y() }, angle1, angle2 ); }
+	void arc_to( const point &center, double radius, double angle1, double angle2 ) { arc_to( center, { center.x() + radius, center.y() }, angle1, angle2 ); }
 
 	void move_by( const point &p ) { move_to( next_point( p ) ); }
 	void line_by( const point &p ) { line_to( next_point( p ) ); }
